+++ HISTORY.rst
@@ -137,6 +137,8 @@
   into a YAML-formatted string.
 - Add `FormMixinFormMeta` to support specifying custom form options in a well
   organized `FormMeta` form class member.
+- Add `get_stored_driver_instance` to allow a file metadata driver instance
+  to have access to its corresponding `StoredDriver` model instance.
 
 4.9.5 (XXXX-XX-XX)
 ==================